env: 패키지 구조 재설정 (#5)
