--- src/main/java/com/umust/dobonglife/course/domain/constant/CourseLevel.java
@@ -1,8 +1,0 @@
-package com.umust.dobonglife.course.domain.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum CourseLevel {
-    BEGINNER, INTERMEDIATE, ADVANCED
-}
