--- src/main/java/com/umust/dobonglife/course/domain/constant/CourseTheme.java
@@ -1,8 +1,0 @@
-package com.umust.dobonglife.course.domain.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum CourseTheme {
-    NATURE, CULTURE, RESTAURANT, HISTORY, FAMILY, ACTIVITY
-}
