+++ src/main/java/com/umust/dobonglife/domain/course/domain/constant/CourseLevel.java
@@ -1,0 +1,8 @@
+package com.umust.dobonglife.domain.course.domain.constant;
+
+import lombok.Getter;
+
+@Getter
+public enum CourseLevel {
+    BEGINNER, INTERMEDIATE, ADVANCED
+}
