+++ src/main/java/com/umust/dobonglife/domain/course/domain/constant/CourseTheme.java
@@ -1,0 +1,8 @@
+package com.umust.dobonglife.domain.course.domain.constant;
+
+import lombok.Getter;
+
+@Getter
+public enum CourseTheme {
+    NATURE, CULTURE, RESTAURANT, HISTORY, FAMILY, ACTIVITY
+}
