+++ src/main/java/com/umust/dobonglife/domain/course/domain/repository/CoursePlansRepository.java
@@ -1,0 +1,13 @@
+package com.umust.dobonglife.domain.course.domain.repository;
+
+import com.umust.dobonglife.domain.course.domain.entity.CoursePlans;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import java.util.List;
+
+public interface CoursePlansRepository extends JpaRepository<CoursePlans, Integer> {
+    @Query("SELECT c FROM Course c WHERE c.id = :courseId")
+    List<CoursePlans> findByCourseIdOrderByDateTime(@Param("courseId") Integer courseId);
+}
