+++ src/main/java/com/umust/dobonglife/domain/course/domain/repository/custom/CourseRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.umust.dobonglife.domain.course.domain.repository.custom;
+
+import com.umust.dobonglife.domain.course.domain.entity.Course;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface CourseRepositoryCustom {
+    Page<Course> findAllRandomOrder(Pageable pageable);
+}
