+++ src/main/java/com/umust/dobonglife/domain/notification/domain/entity/Notification.java
@@ -1,0 +1,4 @@
+package com.umust.dobonglife.domain.notification.domain.entity;
+
+public class Notification {
+}
