+++ src/main/java/com/umust/dobonglife/domain/notification/domain/repository/NotificationRepository.java
@@ -1,0 +1,7 @@
+package com.umust.dobonglife.domain.notification.domain.repository;
+
+import com.umust.dobonglife.domain.notification.domain.entity.Notification;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NotificationRepository extends JpaRepository<Notification, Integer> {
+}
