+++ src/main/java/com/umust/dobonglife/domain/notification/presentation/NotificationController.java
@@ -1,0 +1,4 @@
+package com.umust.dobonglife.domain.notification.presentation;
+
+public class NotificationController {
+}
