+++ src/main/java/com/umust/dobonglife/domain/notification/service/NotificationService.java
@@ -1,0 +1,4 @@
+package com.umust.dobonglife.domain.notification.service;
+
+public class NotificationService {
+}
